mechanics code reorg + small refactoring
