+++ GW2EIEvtcParser/EIData/Mechanics/CastMechanics/Enemy/EnemyCastStartMechanic.cs
@@ -1,0 +1,21 @@
+﻿using GW2EIEvtcParser.ParsedData;
+
+namespace GW2EIEvtcParser.EIData
+{
+
+    internal class EnemyCastStartMechanic : EnemyCastMechanic
+    {
+        protected override long GetTime(AbstractCastEvent evt)
+        {
+            return evt.Time;
+        }
+
+        public EnemyCastStartMechanic(long mechanicID, string inGameName, MechanicPlotlySetting plotlySetting, string shortName, string description, string fullName, int internalCoolDown, CastChecker condition = null) : base(mechanicID, inGameName, plotlySetting, shortName, description, fullName, internalCoolDown, condition)
+        {
+        }
+
+        public EnemyCastStartMechanic(long[] mechanicIDs, string inGameName, MechanicPlotlySetting plotlySetting, string shortName, string description, string fullName, int internalCoolDown, CastChecker condition = null) : base(mechanicIDs, inGameName, plotlySetting, shortName, description, fullName, internalCoolDown, condition)
+        {
+        }
+    }
+}
