+++ GW2EIEvtcParser/EIData/Mechanics/CastMechanics/Friendly/PlayerCastStartMechanic.cs
@@ -1,0 +1,21 @@
+﻿using GW2EIEvtcParser.ParsedData;
+
+namespace GW2EIEvtcParser.EIData
+{
+
+    internal class PlayerCastStartMechanic : PlayerCastMechanic
+    {
+        protected override long GetTime(AbstractCastEvent evt)
+        {
+            return evt.Time;
+        }
+
+        public PlayerCastStartMechanic(long mechanicID, string inGameName, MechanicPlotlySetting plotlySetting, string shortName, string description, string fullName, int internalCoolDown, CastChecker condition = null) : base(mechanicID, inGameName, plotlySetting, shortName, description, fullName, internalCoolDown, condition)
+        {
+        }
+
+        public PlayerCastStartMechanic(long[] mechanicIDs, string inGameName, MechanicPlotlySetting plotlySetting, string shortName, string description, string fullName, int internalCoolDown, CastChecker condition = null) : base(mechanicIDs, inGameName, plotlySetting, shortName, description, fullName, internalCoolDown, condition)
+        {
+        }
+    }
+}
