+++ GW2EIEvtcParser/EIData/Mechanics/EnemyMechanicHelper.cs
