+++ GW2EIEvtcParser/EIData/Mechanics/IDBasedMechanic.cs
@@ -1,5 +1,4 @@
 ﻿using System.Collections.Generic;
-using GW2EIEvtcParser.ParsedData;
 
 namespace GW2EIEvtcParser.EIData
 {
