+++ GW2EIEvtcParser/EIData/Mechanics/SkillMechanics/Enemy/HitMechanics/HitByEnemyMechanic.cs
