+++ GW2EIEvtcParser/EIData/Mechanics/SkillMechanics/Friendly/HitMechanics/FirstHitOnPlayerMechanic.cs
