+++ GW2EIEvtcParser/EIData/Mechanics/SkillMechanics/Friendly/HitMechanics/HitByPlayerMechanic.cs
@@ -1,5 +1,4 @@
-﻿using System.Collections.Generic;
-using GW2EIEvtcParser.ParsedData;
+﻿using GW2EIEvtcParser.ParsedData;
 
 namespace GW2EIEvtcParser.EIData
 {
