+++ GW2EIEvtcParser/EIData/Mechanics/SkillMechanics/Friendly/HitMechanics/HitOnEnemySpeciesMechanic.cs
