+++ GW2EIEvtcParser/EIData/Mechanics/SpawnMechanic.cs
