+++ GW2EIEvtcParser/EIData/Mechanics/StatusMechanics/Friendly/PlayerStatusMechanic.cs
@@ -1,6 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
+﻿using System.Collections.Generic;
 using GW2EIEvtcParser.ParsedData;
 
 namespace GW2EIEvtcParser.EIData
