+++ GW2EIEvtcParser/EIData/Mechanics/StatusMechanics/StatusMechanic.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
 using GW2EIEvtcParser.ParsedData;
 
 namespace GW2EIEvtcParser.EIData
